[FEATURE] Allow other auth providers to use Authorization
Let other providers to use the Authorization header. The Basic
prefix will be used to detect basic authentication when that
header is used. This allows Simple OAuth.

Additionally changes the String class name, since it's not
compatible with PHP7.
