+++ src/Http/Request.php
@@ -7,6 +7,7 @@
 
 namespace Drupal\restful\Http;
 use Drupal\restful\Exception\BadRequestException;
+use Drupal\restful\Util\StringHelper;
 
 /**
  * Deals with everything coming from the consumer.
@@ -432,8 +433,12 @@ class Request implements RequestInterface {
     // for compatibility with Apache PHP CGI/FastCGI.
     // This requires the following line in your ".htaccess"-File:
     // RewriteRule .* - [E=HTTP_AUTHORIZATION:%{HTTP:Authorization}]
-    if (!empty($_SERVER['REDIRECT_HTTP_AUTHORIZATION']) && !isset($username) && !isset($password)) {
-      $authentication = base64_decode(substr($_SERVER['REDIRECT_HTTP_AUTHORIZATION'], 6));
+    $authorization_header = $_SERVER['HTTP_AUTHORIZATION'] ?: $_SERVER['REDIRECT_HTTP_AUTHORIZATION'];
+    if (!empty($authorization_header) && !isset($username) && !isset($password)) {
+      if (!$token = StringHelper::removePrefix('Basic ', $authorization_header)) {
+        return NULL;
+      }
+      $authentication = base64_decode($token);
       list($username, $password) = explode(':', $authentication);
       $_SERVER['PHP_AUTH_USER'] = $username;
       $_SERVER['PHP_AUTH_PW'] = $password;
